Float changes

+++ app/main.py
@@ -13,9 +13,9 @@ class Car:
 class CarWashStation:
     def __init__(
             self,
-            distance_from_city_center: int,
+            distance_from_city_center: float,
             clean_power: int,
-            average_rating: int,
+            average_rating: float,
             count_of_ratings: int
     ) -> None:
         self.distance_from_city_center = distance_from_city_center
